@@ -1,8 +1,7 @@
 module github.com/jfeliu007/goplantuml
 
-go 1.17
+go 1.18
 
-require (
-	github.com/spf13/afero v1.8.2
-	golang.org/x/text v0.3.7 // indirect
-)
+require github.com/spf13/afero v1.8.2
+
+require golang.org/x/text v0.3.7 // indirect
